Move the rest of the internal stuff in `ykcapi/lib.rs` to `ykrt`.
Happily this also allows us to reduce ykcapi's dependencies.
